Refactor/contact (#148)
* transform Contact component into a HomeSection component

* add styling, remove old Contact story

* adjust background

* remove old Contact section files, minor tweak to home page style and layout

* add snapsot test for HomeSection

* add propType for id

* remove unnecessary tests
