--- src/components/contact/Contact.stories.js
@@ -1,6 +1,0 @@
-import React from 'react';
-import { storiesOf } from '@storybook/react';
-
-import Contact from './Contact';
-
-storiesOf('Contact', module).add('renders Contact section', () => <Contact />);
