+++ src/i18n/en.js
@@ -16,15 +16,15 @@ export default function loadEn() {
       'home-about-text':
         'Coding Coach is a free, open-source platform which aims to connect software developers and mentors all over the world.  It is built by a group of talented and passionate developers, designers, engineers, and humans who want to make the engineering world a better place to collaborate.  This project was born out of a desire to provide a platform to connect mentors and mentees throughout the world at no cost.  Coding Coach is created by the people, for the people.',
       'home-about-media-alt': 'Man walking into a laptop',
+      'home-contact-title': 'Contact',
+      'home-contact-text':
+        'We want to hear your thoughts! Feel free to join our {{slackOrg}} or send us an email at {{email}}',
+      'home-contact-media-alt':
+        'Person using a magnifying glass to see the details of another guy.',
       'home-mission-title': 'Mission',
       'home-mission-text':
         'We believe that mentorship should be accessible to all people regardless of location or financial standing. In pursuit of this goal we will provide a free and open source platform to facilitate mentorship connections. Our mission is to foster a greater sense of collaboration and inclusiveness in the technical industry by providing a platform to aid the mentorship process.',
       'home-mission-media-alt': 'Group of people',
-      'contact-page-header': 'Contact',
-      'contact-page-msg':
-        'We want to hear your thoughts! Feel free to join our {{slackOrg}} or send us an email at {{email}}',
-      'contact-page-media-alt':
-        'Person using a magnifying glass to see the details of another guy.',
       'slack-org': 'Slack Organization',
     },
   };
