+++ src/pages/home/__tests__/SocialMedia.test.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { render } from 'react-testing-library';
-import SocialMedia from '../SocialMedia';
+import SocialMedia from '../components/SocialMedia';
 import config from 'config/constants';
 
-test('renders the contact page', () => {
+test('renders the social media icons', () => {
   const { container } = render(<SocialMedia />);
 
   expect(container.firstChild).toMatchSnapshot();
